Tenpai checker.
The idea of covering general mahjong tenpai / mahjong hands is simple:

(exclude some stupid MCR hands)
Check Kokushimusou (13 orphans)
Check 7 pairs
Check classic 13 orphans, i.e. 13 completely separated cards (this part was not coded, but it is easy)
Check standard mahjong hand of 3 blocks and a pair/eye

Opened blocks are not coded, as by definition those would need not be fed into checking a hand if mahjong or not.

The logic is simple. Starting in ascending order, check if a pair exist. If there is a pair, copy the hand, remove the pair. Feed the 3x block into the function mahjong checker. Check if it is triplet or run, if yes, then feed the rest of the block into the same function. If not, return 0 and increase the pair count, Look for other pairs, and then rinse and repeat.

If in the mahjong checker the hand size is ever zero, it means the whole hand is checked and the function would loop back up into level 0 and return 1, which is mahjong.

Otherwise return 0, the hand is not mahjong.

Allowing jokers and read tenpai / 1-card-to-tenpai etc is just reserve engineering allowing n^k of combinations testing into the hand.